Remove old RP.
